Expose painter builders in CoilImage
